Break React components into their own files
